requirements.txt
streamlit
snowflake-ml-python
snowflake-snowpark-python
tabulate
sqlalchemy
psycopg2-binary
openai
numpy
pandas
